requirements.txt
Django==1.11.15
pytz==2018.5
requests==2.9.1
urllib3==1.13.1
virtualenv==16.0.0
djangorestframework
djangorestframework-jwt